add summary modal
