+++ src/assets/javascript/main.js
@@ -2,7 +2,7 @@
 import './interactions/copy.js';
 import './interactions/forms.js';
 import './interactions/header.js';
-import './interactions/modal-signup.js';
+import './interactions/modals.js';
 
 // Animations
 import './animations/blog-post.js';
