+++ src/assets/sass/custom/_misc.scss
@@ -20,6 +20,11 @@
 .scrollbar\:dark::-webkit-scrollbar-thumb { background: rgba(var(--edge-1-i), 1); border-radius: 999px; border:4px solid rgba(var(--sheet-4-i), 1); }
 .scrollbar\:dark::-webkit-scrollbar-track { background: rgba(var(--sheet-4-i), 1); }
 
+.scrollbar\:light { scrollbar-color: rgba(var(--edge-1), 1) rgba(var(--sheet-4), 1); }
+.scrollbar\:light::-webkit-scrollbar { width: 1rem; }
+.scrollbar\:light::-webkit-scrollbar-thumb { background: rgba(var(--edge-1), 1); border-radius: 999px; border:4px solid rgba(var(--sheet-4), 1); }
+.scrollbar\:light::-webkit-scrollbar-track { background: rgba(var(--sheet-4), 1); }
+
 .gradient-text { background-clip: text; -webkit-background-clip: text; -webkit-text-fill-color: transparent; }
 
 // Homepage Phone Scroll
